requirements.txt
# Flask web framework
flask==3.0.0
flask-cors==4.0.0

# AI/LLM SDKs
openai==1.54.0
anthropic==0.18.1
instructor>=1.0.0
pydantic>=2.0.0

# HTTP client (pin compatible version for openai)
httpx==0.27.2

# Environment variables
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.25

# Utilities
requests==2.31.0

# Production server
gunicorn==21.2.0

# WhatsApp/SMS integration
twilio==9.0.4
